Index

ConstrainedSystems.IFHERK — Type

IFHERK(u,f,Δt,plan_intfact,B₁ᵀ,B₂,r₁,r₂;[tol=1e-3],[issymmetric=false],[rk::RKParams=RK31])

Construct an integrator to advance a system of the form

du/dt - Au = -B₁ᵀf + r₁(u,t) B₂u = r₂(u,t)

The resulting integrator will advance the system (u,f) by one time step, Δt. The optional argument tol sets the tolerance of iterative saddle-point solution, if applicable.

Arguments

u : example of state vector data

f : example of constraint force vector data

Δt : time-step size

plan_intfact : constructor to set up integrating factor operator for A that will act on type u (by left multiplication) and return same type as u

plan_constraints : constructor to set up the

B₁ᵀ : operator acting on type f and returning type u

B₂ : operator acting on type u and returning type f

r₁ : operator acting on type u and t and returning u

r₂ : operator acting on type u and t and returning type f

ConstrainedSystems.IFRK — Type

IFRK(u,Δt,plan_intfact,r₁;[rk::RKParams=RK31])

Construct an integrator to advance a system of the form

du/dt - Au = r₁(u,t)

The resulting integrator will advance the state u by one time step, Δt.

Arguments

u : example of state vector data

Δt : time-step size

plan_intfact : constructor to set up integrating factor operator for A that will act on type u (by left multiplication) and return same type as u

r₁ : operator acting on type u and t and returning u

ConstrainedSystems.RK — Type

RK(u,Δt,r₁;[rk::RKParams=RK31])

Construct an integrator to advance a system of the form

du/dt = r₁(u,t)

The resulting integrator will advance the state u by one time step, Δt.

Arguments

u : example of state vector data

Δt : time-step size

r₁ : operator acting on type u and t and returning u

ConstrainedSystems.SaddleSystem — Method

SaddleSystem

Construct a saddle-point system operator from the constituent operator blocks. The resulting object can be used with * and \ to multiply and solve. The saddle-point problem has the form

$\begin{bmatrix}A & B_1^T \\ B_2 & C \end{bmatrix} \begin{pmatrix} u \\ f \end{pmatrix} = \begin{pmatrix} r_1 \\ r_2 \end{pmatrix}$

Constructors

SaddleSystem(A::AbstractMatrix,B₂::AbstractMatrix,B₁ᵀ::AbstractMatrix,C::AbstractMatrix[,eltype=Float64]). Blocks are given as matrices. Must have consistent sizes to stack appropriately. If this is called with SaddleSystem(A,B₂,B₁ᵀ), it sets C to zero automatically.

SaddleSystem(A,B₂,B₁ᵀ,C,u,f[,eltype=Float64]). Operators A, B₂, B₁ᵀ, C are given in various forms, including matrices, functions, and function-like objects. u and f are examples of the data types in the corresponding solution and right-hand side vectors. Guidelines:

The entries A and B₂ must be able to act upon u (either by multiplication or as a function) and B₁ᵀ and C must be able to act on f (also, either by multiplication or as a function).

A and B₁ᵀ should return data of type u, and B₂ and C should return data of type f.

A must be invertible and be outfitted with operators `andldiv!`.

Both u and f must be subtypes of AbstractArray: they must be equipped with size and vec functions and with a constructor of the form T(data) where T is the data type of u or f and data is the wrapped data array.

If called as SaddleSystem(A,B₂,B₁ᵀ,u,f), the C block is omitted and assumed to be zero.

If called with SaddleSystem(A,u), this is equivalent to calling SaddleSystem(A,nothing,nothing,u,[]), then this reverts to the unconstrained system described by operator A.

The list of vectors u and f in any of these constructors can be bundled together as a SaddleVector, e.g. SaddleSystem(A,B₂,B₁ᵀ,SaddleVector(u,f)).

ConstrainedSystems.SaddleVector — Type

SaddleVector(u,f)

Construct a vector of a state part u and constraint part f of a saddle-point vector, to be associated with a SaddleSystem.

ConstrainedSystems.System — Type

Abstract type for a system of ODEs

Base.eltype — Method

Base.eltype(::SaddleSystem)

Report the element type of a SaddleSystem.

Base.size — Method

Base.size(::SaddleSystem)

Report the size of a SaddleSystem.

ConstrainedSystems.constraint — Method

constraint(x::SaddleVector)

Provide the constraint part of the given saddle vector x

ConstrainedSystems.state — Method

state(x::SaddleVector)

Provide the state part of the given saddle vector x